+++ tests/test_validation.py
@@ -308,6 +308,12 @@ class TestEmail(ValidationBase, unittest.TestCase):
 
 
 class TestEvent(ValidationBase, unittest.TestCase):
+    def test_party(self) -> None:
+        for good in ("agent", "customer"):
+            self.check_transaction({"event": {"party": good}})
+        for bad in ("bad", 1, ""):
+            self.check_invalid_transaction({"event": {"party": bad}})
+
     def test_transaction(self) -> None:
         self.check_transaction_str_type("event", "transaction_id")
 
@@ -324,7 +330,9 @@ class TestEvent(ValidationBase, unittest.TestCase):
         for good in (
             "account_creation",
             "account_login",
+            "credit_application",
             "email_change",
+            "fund_transfer",
             "password_reset",
             "payout_change",
             "purchase",
@@ -370,6 +378,23 @@ class TestOrder(ValidationBase, unittest.TestCase):
 
 
 class TestPayment(ValidationBase, unittest.TestCase):
+    def test_method(self) -> None:
+        for good in (
+            "bank_debit",
+            "bank_redirect",
+            "bank_transfer",
+            "buy_now_pay_later",
+            "card",
+            "crypto",
+            "digital_wallet",
+            "gift_card",
+            "real_time_payment",
+            "rewards",
+        ):
+            self.check_transaction({"payment": {"method": good}})
+        for bad in ("bad", 1, ""):
+            self.check_invalid_transaction({"payment": {"method": bad}})
+
     def test_processor(self) -> None:
         for good in ("adyen", "stripe"):
             self.check_transaction({"payment": {"processor": good}})