+++ HISTORY.rst
@@ -9,6 +9,12 @@ History
 * Setuptools has been replaced with the uv build backend for building the
   package.
 * Added ``securepay`` to the ``/payment/processor`` validation.
+* Added ``credit_application`` and ``fund_transfer`` to the ``/event/type``
+  validation.
+* Added the input ``/event/party``. This is the party submitting the
+  transaction.
+* Added the input ``/payment/method``. This is the payment method associated
+  with the transaction.
 
 3.1.0 (2025-05-23)
 ++++++++++++++++++